fix some error msgs. attention! clear role_permissions table and re-seeder it.

+++ app/Http/Controllers/Traits/HasDelete.php
@@ -2,12 +2,17 @@
 
 namespace App\Http\Controllers\Traits;
 
+use App\Http\Controllers\RoleController;
 use Illuminate\Http\RedirectResponse;
 
 trait HasDelete
 {
     public function delete($id): RedirectResponse
     {
+        $authorization = RoleController::can($this->delete_permission);
+        if ($authorization !== null) {
+            return $authorization;
+        }
         if ($this->model::where('id', '=', $id)->delete()) {
             $this->cascadeDelete($id);
             return back()->with('success', 'deleted successfully');
